requirements.txt
psycopg2-binary
boto3
pandas
datetime
yfinance
tweepy
textblob
re